cubit/view naming convention and state cubit logic separation
