+++ lib/onboarding/onboarding.dart
@@ -4,3 +4,5 @@ export 'third/onboarding_third.dart';
 export 'tos/onboarding_terms.dart';
 export 'gen_phrase/onboarding_gen_phrase.dart';
 export 'recovery/onboarding_recovery.dart';
+export 'submit_enterprise_user/submit_enterprise_user.dart';
+export 'wallet_type/choose_wallet_type.dart';
