+++ lib/onboarding/submit_enterprise_user/cubit/submit_enterprise_user_state.dart
@@ -1,4 +1,3 @@
-import 'package:bloc/bloc.dart';
 import 'package:file_picker/file_picker.dart';
 
 class SubmitEnterpriseUserState {
@@ -10,13 +9,5 @@ class SubmitEnterpriseUserState {
   SubmitEnterpriseUserState copyWith({PlatformFile? rsaFile, String? did}) {
     return SubmitEnterpriseUserState(
         rsaFile: rsaFile ?? this.rsaFile, did: did ?? this.did);
-  }
-}
-
-class SubmitEnterpriseUserCubit extends Cubit<SubmitEnterpriseUserState> {
-  SubmitEnterpriseUserCubit() : super(SubmitEnterpriseUserState());
-
-  void setRSAFile(PlatformFile? rsaFile) {
-    emit(state.copyWith(rsaFile: rsaFile));
   }
 }
