+++ lib/self_issued_credential/cubit/self_issued_credential_state.dart
@@ -1,0 +1,17 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'self_issued_credential_state.freezed.dart';
+
+@freezed
+class SelfIssuedCredentialState with _$SelfIssuedCredentialState {
+  const factory SelfIssuedCredentialState.initial() = Initial;
+
+  const factory SelfIssuedCredentialState.loading() = Loading;
+
+  const factory SelfIssuedCredentialState.error(String message) = Error;
+
+  const factory SelfIssuedCredentialState.warning(String message) = Warning;
+
+  const factory SelfIssuedCredentialState.credentialCreated() =
+  CredentialCreated;
+}
