+++ lib/self_issued_credential/sef_issued_credential.dart
@@ -1,0 +1,6 @@
+export 'cubit/self_issued_credential_cubit.dart';
+export 'cubit/self_issued_credential_state.dart';
+export 'models/self_issued.dart';
+export 'models/self_issued_credential.dart';
+export 'view/models/self_issued_credential_model.dart';
+export 'view/sef_issued_credential_button.dart';
