+++ lib/self_issued_credential/view/sef_issued_credential_button.dart
@@ -5,46 +5,11 @@ import 'package:talao/l10n/l10n.dart';
 import 'package:talao/self_issued_credential/cubit/self_issued_credential_cubit.dart';
 import 'package:talao/wallet/cubit/wallet_cubit.dart';
 
+import '../cubit/self_issued_credential_state.dart';
+import 'models/self_issued_credential_model.dart';
+
 typedef SelfIssuedCredentialButtonClick = SelfIssuedCredentialDataModel
     Function();
-
-class SelfIssuedCredentialDataModel {
-  final String? givenName;
-  final String? familyName;
-  final String? telephone;
-  final String? email;
-  final String? address;
-  final String? companyName;
-  final String? companyWebsite;
-  final String? jobTitle;
-
-  SelfIssuedCredentialDataModel({
-    this.givenName,
-    this.familyName,
-    this.telephone,
-    this.email,
-    this.address,
-    this.companyName,
-    this.companyWebsite,
-    this.jobTitle,
-  });
-
-  @override
-  String toString() {
-    return '''
-    SelfIssuedCredentialDataModel {
-                  givenName : $givenName,
-                  familyName : $familyName,
-                  telephone : $telephone,
-                  email : $email,
-                  address : $address,
-                  companyName : $companyName,
-                  companyWebsite : $companyWebsite,
-                  jobTitle : $jobTitle,
-    }
-    ''';
-  }
-}
 
 class SelfIssuedCredentialButton extends StatelessWidget {
   final SelfIssuedCredentialButtonClick selfIssuedCredentialButtonClick;
